+++ modulo3/projeto-labex/src/hooks/useProtectedPage.js
@@ -1,0 +1,14 @@
+import { useEffect } from 'react'
+import { useNavigate } from 'react-router-dom'
+import { goToLoginPage } from '../routes/coordinator'
+
+export const useProtectedPage = () => {
+    const navigate = useNavigate()
+
+    useEffect(() => {
+        const token = localStorage.getItem('token')
+        if(token === null) {
+            goToLoginPage(navigate)
+        }
+    })
+}
