+++ modulo3/projeto-labex/src/pages/CreateTripPage/CreateTripPage.js
@@ -1,6 +1,9 @@
 import React from 'react'
+import { useProtectedPage } from '../../hooks/useProtectedPage'
 
 export default function CreateTripPage() {
+    useProtectedPage()
+
     return <div>
         Criação de viagem
     </div>
